refactor: place unordered list in a separate folder
